requirements.txt
APScheduler==3.6.3
asttokens==2.0.8
attrs==19.3.0
Automat==0.8.0
autopep8==1.7.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.4
cachetools==4.2.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==22.3.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
debugpy==1.6.3
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info==0.23ubuntu1
entrypoints==0.3
et-xmlfile==1.1.0
executing==1.1.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib-metadata==1.5.0
incremental==16.10.1
ipykernel==6.16.0
ipython==8.5.0
jedi==0.18.1
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==7.3.5
jupyter-core==4.11.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
MarkupSafe==1.1.0
matplotlib-inline==0.1.6
more-itertools==4.2.0
moviepy==1.0.3
nest-asyncio==1.5.5
netifaces==0.10.4
numpy==1.23.3
oauthlib==3.1.0
olefile==0.46
openai==0.25.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
pandas-stubs==1.5.2.221213
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
pika==1.3.0
Pillow==9.2.0
pip==20.0.2
proglog==0.1.10
prompt-toolkit==3.0.31
protobuf==3.6.1
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycodestyle==2.9.1
pycups==1.9.73
Pygments==2.13.0
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian==0.1.36ubuntu1
python-dotenv==0.21.0
python-telegram-bot==13.15
pytube==12.1.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==24.0.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
sos==4.4
ssh-import-id==5.10
stack-data==0.5.1
systemd-python==234
toml==0.10.2
tornado==6.1
tqdm==4.64.1
traitlets==5.4.0
Twisted==18.9.0
types-pytz==2022.7.0.0
typing-extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wcwidth==0.2.5
wheel==0.34.2
xkit==0.0.0
zipp==1.0.0
zope.interface==4.7.1
